refactor: 💡 Improvement code [SCRIPTS]
