+++ src/index.js
@@ -1,12 +1,12 @@
 // SCRIPTS [IMPORTANTS]
-import "@scripts/bootstrap.min.js";
-import "@scripts/particles.js";
-import "@scripts/google-analytics.js";
+import "@scripts/vendors/bootstrap.min.js";
+import "@scripts/vendors/particles.js";
+import "@scripts/vendors/google-analytics.js";
 
 // SCRIPTS [ANIMATIONS]
-import "@scripts/animate-back-to-top.js";
-import "@scripts/animate-counter.js";
-import "@scripts/animate-marquee.js";
+import "@scripts/animations/back-to-top.js";
+import "@scripts/animations/counter.js";
+import "@scripts/animations/marquee.js";
 
 // SCSS
 import "@scss/index.scss";
