+++ src/scripts/animations/back-to-top.js
