+++ src/scripts/animations/counter.js
