+++ src/scripts/animations/marquee.js
